test: add prod/staging monitoring
