+++ .github/workflows/e2e.yml
@@ -5,18 +5,6 @@ on:
     branches: [main]
   push:
     branches: [main]
-  workflow_dispatch:
-    # TODO: implement
-    # inputs:
-    #   environment:
-    #     description: "Choose the environment"
-    #     required: true
-    #     default: "production"
-    #     type: choice
-    #     options:
-    #       - production
-    #       - staging
-  # TODO: add cronjob
 
 jobs:
   test:
@@ -121,7 +109,7 @@ jobs:
           NEXT_PUBLIC_SUPABASE_ANON_KEY: ${{ secrets.NEXT_PUBLIC_SUPABASE_ANON_KEY }}
           NEXT_PUBLIC_HUBSPOT_PORTAL_ID: ${{ secrets.NEXT_PUBLIC_HUBSPOT_PORTAL_ID }}
           NEXT_PUBLIC_HUBSPOT_FORM_ID: ${{ secrets.NEXT_PUBLIC_HUBSPOT_FORM_ID }}
-        run: pnpm --filter app-e2e test
+        run: pnpm --filter app-e2e test:dev
         continue-on-error: true
 
       - name: Upload test screenshots
