+++ apps/app-e2e/playwright.config.prod.ts
@@ -4,9 +4,13 @@ import { defineConfig, devices } from "@playwright/test";
  * Read environment variables from file.
  * https://github.com/motdotla/dotenv
  */
-// import dotenv from 'dotenv';
-// import path from 'path';
-// dotenv.config({ path: path.resolve(__dirname, '.env') });
+const APP_URL = process.env.TEST_APP_URL;
+
+if (!APP_URL) {
+  console.error(
+    "Error: TEST_APP_URL environment variable is required for production tests.",
+  );
+}
 
 /**
  * See https://playwright.dev/docs/test-configuration.
@@ -26,8 +30,7 @@ export default defineConfig({
   /* Shared settings for all the projects below. See https://playwright.dev/docs/api/class-testoptions. */
   use: {
     /* Base URL to use in actions like `await page.goto('/')`. */
-    baseURL: "http://localhost:3000",
-
+    baseURL: APP_URL,
     /* Collect trace when retrying the failed test. See https://playwright.dev/docs/trace-viewer */
     trace: "on-first-retry",
   },
@@ -35,7 +38,8 @@ export default defineConfig({
   /* Configure projects for major browsers */
   projects: [
     {
-      name: "Google Chrome",
+      name: "monitor",
+      testMatch: /.*\.prod\.spec\.ts/,
       use: {
         ...devices["Desktop Chrome"],
         channel: "chrome",
@@ -49,11 +53,4 @@ export default defineConfig({
       },
     },
   ],
-
-  /* Run your local dev server before starting the tests */
-  webServer: {
-    command: "pnpm --filter app dev:mock",
-    url: "http://127.0.0.1:3000",
-    reuseExistingServer: !process.env.CI,
-  },
 });
